remove redundant cast
git-svn-id: svn+ssh://developer.mitre.org/svnroot/giscore/trunk@710 5c99b798-f2dd-11dd-9a12-8f84353f4b73

+++ giscore/src/org/mitre/giscore/output/dbf/DbfOutputStream.java
@@ -187,7 +187,7 @@ public class DbfOutputStream implements IGISOutputStream, IDbfConstants {
 			stream.writeInt((int) buffer.count(), ByteOrder.LITTLE_ENDIAN);
 			stream.writeShort((short) ((schema.getKeys().size() * 32) + 33),
 					ByteOrder.LITTLE_ENDIAN);
-			stream.writeShort((short) getRecordLength(), ByteOrder.LITTLE_ENDIAN);
+			stream.writeShort( getRecordLength(), ByteOrder.LITTLE_ENDIAN);
 
 			// Fill in reserved and unused header fields we don't care about
 			// with
